ci: lint the Windows build and clear its findings
CI linted only the linux build, so Windows-only files (named-pipe IPC,
Job Object hooks, registry native host) were never linted. The lint job
is now a matrix over GOOS linux and windows.

The windows run found five issues, fixed here:
- ownershipsnapshot: read FILE_BASIC_INFO's times as uint64; changeTime
  is only compared for equality, so no int64 -> uint64 conversion (G115).
- hook: a Windows pid is the DWORD CreateProcess returned (narrow G115
  nolint with that reason).
- cli, nativehost: two os.Stat calls on paths under the user's own
  LOCALAPPDATA and config dir (narrow G703 nosec with the reason).
- config, ipc tests: an unchecked CloseHandle, and a %v on an error that
  now wraps with %w.

+++ internal/ownershipsnapshot/fileid_windows.go
@@ -14,11 +14,14 @@ func openSource(path string) (*os.File, error) {
 	return os.Open(path)
 }
 
+// fileBasicInfo mirrors FILE_BASIC_INFO. The times are LARGE_INTEGERs, read
+// here as uint64 because papio only compares changeTime for equality as part of
+// a file identity; the bits are the same either way.
 type fileBasicInfo struct {
-	creationTime   int64
-	lastAccessTime int64
-	lastWriteTime  int64
-	changeTime     int64
+	creationTime   uint64
+	lastAccessTime uint64
+	lastWriteTime  uint64
+	changeTime     uint64
 	fileAttributes uint32
 }
 
@@ -39,7 +42,7 @@ func fileIdentity(file *os.File, _ os.FileInfo) fileID {
 	return fileID{
 		device:  uint64(handleInfo.VolumeSerialNumber),
 		inode:   uint64(handleInfo.FileIndexHigh)<<32 | uint64(handleInfo.FileIndexLow),
-		changeA: uint64(basic.changeTime),
+		changeA: basic.changeTime,
 		known:   true,
 	}
 }
